Auth as one function and use common hash function
