+++ auth_test.go
@@ -6,26 +6,23 @@ import (
 )
 
 func TestBadAuth(t *testing.T) {
-	r, _ := http.NewRequest("GET", "/auth/basic/bad", nil)
-	rr := executeRequest(r, BasicAuth(Hello, "user", "user"))
+	r, _ := http.NewRequest("GET", "/auth/basic/", nil)
+	r.SetBasicAuth("user", "user")
+	rr := executeRequest(r, DynamicAuth)
 	checkResponse(t, rr, http.StatusUnauthorized)
 }
 
 func TestGoodAuth(t *testing.T) {
-	r, _ := http.NewRequest("GET", "/auth/basic/bad", nil)
+	r, _ := http.NewRequest("GET", "/auth/basic/", nil)
 	r.SetBasicAuth("admin", "admin")
-	rr := executeRequest(r, BasicAuth(Hello, "admin", "admin"))
+	rr := executeRequest(r, DynamicAuth)
 	checkResponse(t, rr, http.StatusOK)
 }
 
 func TestBadDynamicAuth(t *testing.T) {
-	r, _ := http.NewRequest("GET", "/auth/basic/user/password", nil)
+	r, _ := http.NewRequest("GET", "/auth/basic/user/bad", nil)
 	r.SetBasicAuth("user", "failed")
-	rr := executeVarsRequest("/auth/basic/{user}/{password}", r, DynamicAuth)
-	checkResponse(t, rr, http.StatusBadRequest)
-	r, _ = http.NewRequest("GET", "/auth/basic/user/bad", nil)
-	r.SetBasicAuth("user", "failed")
-	rr = executeVarsRequest("/auth/basic/{username}/{password}", r, DynamicAuth)
+	rr := executeVarsRequest("/auth/basic/{username}/{password}", r, DynamicAuth)
 	checkResponse(t, rr, http.StatusUnauthorized)
 }
