+++ go.mod
@@ -3,7 +3,7 @@ module github.com/Cyb3r-Jak3/simple_http_server/v2
 go 1.16
 
 require (
-	github.com/Cyb3r-Jak3/common v0.1.1
+	github.com/Cyb3r-Jak3/common v0.1.2
 	github.com/brianvoe/gofakeit/v6 v6.5.0
 	github.com/gorilla/mux v1.8.0
 )
